refactor(docs): move skillCodeBlockPlugin to docs/vite-plugins/
Review feedback on #3501.
